# TaskRonaut/backend/src/config/swaggerAll.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Taskronaut API",
    "version": "1.0.0",
    "description": "API documentation for Taskronaut"
  },
...
  "paths": {
...
    "/boards": {
      "get": {
        "summary": "Liefert alle Boards als JSON zurück",
        "tags": [
          "Board"
        ],
        "description": "Gibt eine Liste aller Boards zurück.",
        "responses": {
          "200": {
            "description": "Eine Liste von Boards",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "boardId": {
                        "type": "integer",
                        "description": "Die ID des Boards"
                      },
                      "boardName": {
                        "type": "string",
                        "description": "Der Name des Boards"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Erstellt ein neues Board",
        "tags": [
          "Board"
        ],
        "requestBody": {
          "description": "Name des neuen Boards",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "boardName": {
                    "type": "string",
                    "description": "Der Name des Boards"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Board erfolgreich erstellt"
          },
          "400": {
            "description": "Fehler beim Erstellen des Boards"
          }
        }
      }
    },
    "/boards/{id}": {
      "delete": {
        "summary": "Löscht ein existierendes Board",
        "tags": [
          "Board"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID des zu löschenden Boards"
          }
        ],
        "responses": {
          "200": {
            "description": "Board erfolgreich entfernt"
          },
          "404": {
            "description": "Board nicht gefunden"
          }
        }
      }
    },
    "/board-members": {
      "get": {
        "summary": "Gibt alle Board-Mitglieder zurück",
        "tags": [
          "BoardMember"
        ],
        "description": "Liefert eine Liste aller Mitglieder eines Boards.",
        "responses": {
          "200": {
            "description": "Eine Liste von Board-Mitgliedern",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "userId": {
                        "type": "integer",
                        "description": "Die ID des Mitglieds"
                      },
                      "boardId": {
                        "type": "integer",
                        "description": "Die ID des Boards"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Fügt ein Mitglied einem Board hinzu",
        "tags": [
          "BoardMember"
        ],
        "requestBody": {
          "description": "ID des Users und ID des Boards",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "integer",
                    "description": "Die ID des Users"
                  },
                  "boardId": {
                    "type": "integer",
                    "description": "Die ID des Boards"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mitglied erfolgreich hinzugefügt"
          },
          "400": {
            "description": "Fehler beim Hinzufügen des Mitglieds"
          }
        }
      }
    },
    "/board-members/{id}": {
      "delete": {
        "summary": "Entfernt ein Mitglied von einem Board",
        "tags": [
          "BoardMember"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID des Mitglieds"
          }
        ],
        "responses": {
          "200": {
            "description": "Mitglied erfolgreich entfernt"
          },
          "404": {
            "description": "Mitglied nicht gefunden"
          }
        }
      }
    },
    "/lists": {
      "get": {
        "summary": "Liefert alle Listen zurück",
        "description": "Gibt eine Liste aller vorhandenen Listen aus der Datenbank zurück.",
        "tags": [
          "List"
        ],
        "responses": {
          "200": {
            "description": "Eine Liste von Listen",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "listID": {
                        "type": "integer",
                        "description": "Die eindeutige ID der Liste"
                      },
                      "listName": {
                        "type": "string",
                        "description": "Der Name der Liste"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Fehler beim Zugriff auf die Datenbank"
          }
        }
      },
      "post": {
        "summary": "Erstellt eine neue Liste",
        "description": "Fügt eine neue Liste in die Datenbank hinzu.",
        "tags": [
          "List"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "listName": {
                    "type": "string",
                    "description": "Der Name der neuen Liste",
                    "example": "Meine Aufgabenliste"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Liste erfolgreich erstellt"
          },
          "400": {
            "description": "Fehler beim Erstellen der Liste"
          }
        }
      }
    },
    "/lists/{id}": {
      "delete": {
        "summary": "Löscht eine Liste",
        "description": "Löscht eine Liste basierend auf der übergebenen ID.",
        "tags": [
          "List"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID der zu löschenden Liste"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste erfolgreich gelöscht"
          },
          "404": {
            "description": "Liste nicht gefunden"
          },
          "500": {
            "description": "Fehler beim Löschen der Liste"
          }
        }
      }
    },
...
  }
}